hotfix: redirecionamento de tracker de refeições para a rota de tracker
