+++ src/apps/accounts/ui_views.py
@@ -2,9 +2,6 @@ from django.contrib.auth.decorators import login_required
 from django.shortcuts import render
 from django.views.decorators.http import require_GET
 
-from apps.foods.models import Food
-from apps.tracker.models import Meal
-
 
 @require_GET
 def register_page(request):
@@ -45,28 +42,3 @@ def password_reset_confirm_page(request, token: str):
 @require_GET
 def password_reset_success_page(request):
     return render(request, "accounts/password_reset_success.html")
-
-
-@require_GET
-@login_required
-def nutritional_profile_page(request):
-    return render(request, "accounts/nutritional_profile.html")
-
-
-@require_GET
-@login_required
-def tracker_page(request):
-    foods = Food.objects.all().order_by("name")
-    meals = (
-        Meal.objects.filter(user=request.user)
-        .prefetch_related("items__food")
-        .order_by("-eaten_at")
-    )
-    return render(
-        request,
-        "accounts/tracker.html",
-        {
-            "foods": foods,
-            "meals": meals,
-        },
-    )
