+++ src/apps/tracker/ui_urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+
+from . import ui_views
+
+app_name = "tracker-ui"
+
+urlpatterns = [
+    path("", ui_views.tracker_page, name="tracker"),
+]
