+++ src/core/urls.py
@@ -4,7 +4,7 @@ from django.urls import include, path
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("accounts/", include("apps.accounts.ui_urls")),
-    path("foods/", include("apps.foods.ui_urls")),
+    path("tracker/", include("apps.tracker.ui_urls")),
     path("profiles/", include("apps.profiles.ui_urls")),
     path("api/accounts/", include("apps.accounts.urls")),
     path("api/profiles/", include("apps.profiles.urls")),
